// projects/zweihander-1e/enums.ts
export enum TalentId {
  AETHEREAL_ALIGNMENT = 'talent:aethereal_alignment',
  AMBIDEXTERITY = 'talent:ambidexterity',
  APPALLING_MIEN = 'talent:appalling_mien',
  ARBALESTS_SPEED = 'talent:arbalests_speed',
  AZIMUTH = 'talent:azimuth',
  BAD_AXX = 'talent:bad_axx',
  BATTLE_MAGICK = 'talent:battle_magick',
  BEATDOWN = 'talent:beatdown',
  BLOOD_MAGICK = 'talent:blood_magick',
  CAROUSING = 'talent:carousing',
  CHEAP_SHOT = 'talent:cheap_shot',
  CLINCH_FIGHTER = 'talent:clinch_fighter',
  DETERMINATION = 'talent:determination',
  DIE_HARD = 'talent:die_hard',
  DOPPELGANGER = 'talent:doppelganger',
  EAGLE_EYES = 'talent:eagle_eyes',
  FENCERS_PANACHE = 'talent:fencers_panache',
  FORKED_TONGUE = 'talent:forked_tongue',
  GALLOWS_HUMOR = 'talent:gallows_humor',
  GANGSTER_GRIP = 'talent:gangster_grip',
  GATECRASHER = 'talent:gatecrasher',
  GROUND_POUND = 'talent:ground_pound',
  GRUESOME_SHOT = 'talent:gruesome_shot',
  HANDSPRING = 'talent:handspring',
  HARD_TO_KILL = 'talent:hard_to_kill',
  HIGHER_MYSTERIES = 'talent:higher_mysteries',
  HOLDOUT = 'talent:holdout',
  HOUSEBREAKER = 'talent:housebreaker',
  IMPENETRABLE_WALL = 'talent:impenetrable_wall',
  IMPERVIOUS_MIND = 'talent:impervious_mind',
  INDIFFERENCE = 'talent:indifference',
  INCREDIBLE_NUMERATION = 'talent:incredible_numeration',
  INSTINCTS = 'talent:instincts',
  KIDNEY_SHOT = 'talent:kidney_shot',
  KNIFEWORK = 'talent:knifework',
  LARCENY = 'talent:larceny',
  LEFT_HANDED_PATH = 'talent:left_handed_path',
  LIGHTNING_REACTION = 'talent:lightning_reaction',
  LIGHT_SLEEPER = 'talent:light_sleeper',
  LONG_WINDED = 'talent:long_winded',
  MARINER = 'talent:mariner',
  MENACING_DEMEANOR = 'talent:menacing_demeanor',
  MEETING_OF_THE_MINDS = 'talent:meeting_of_the_minds',
  MILITARY_FORMATION = 'talent:military_formation',
  MINE_CRAFT = 'talent:mine_craft',
  MOUNTED_DEFENSE = 'talent:mounted_defense',
  MULTILINGUAL = 'talent:multilingual',
  NERVES_OF_STEEL = 'talent:nerves_of_steel',
  NO_MERCY = 'talent:no_mercy',
  OVERWHELMING_FORCE = 'talent:overwhelming_force',
  RUN_AMOK = 'talent:run_amok',
  RURAL_SENSIBILITY = 'talent:rural_sensibility',
  SADDLE_BORN = 'talent:saddle_born',
  SECOND_SKIN = 'talent:second_skin',
  SECRET_SIGNS = 'talent:secret_signs',
  SHIELD_SLAM = 'talent:shield_slam',
  SHOOT_FROM_THE_HIP = 'talent:shoot_from_the_hip',
  SIEGECRAFT = 'talent:siegecraft',
  SILVER_TONGUE = 'talent:silver_tongue',
  SPIRITED_CHARGE = 'talent:spirited_charge',
  SPRINT = 'talent:sprint',
  STRANGLERS_UNION = 'talent:stranglers_union',
  STREETWISE = 'talent:streetwise',
  STRONG_JAW = 'talent:strong_jaw',
  SUPERNATURAL_PARANOIA = 'talent:supernatural_paranoia',
  SWORD_BOARD = 'talent:sword_board',
  TAKE_EM_DOWN = 'talent:take_em_down',
  THERE_WILL_BE_BLOOD = 'talent:there_will_be_blood',
  TOUGH_AS_NAILS = 'talent:tough_as_nails',
  TRUE_GRIT = 'talent:true_grit',
  WINDS_OF_CHANGE = 'talent:winds_of_change',
  WORLDLY = 'talent:worldly'
}